minor: move createEntitysetDefinition out
